new CWTrie a pure Objective-C implementation of Tries
The idea of this is to write an implementation of tries in pure
Objective-C and then refine it and optimize it. Just wanted to start in
pure Objective-C and see how well that performs and then do things like
drop down to pure C if it gives good performance boosts.